fix styling and add tags to container
